+++ apps/examples/src/pages/change-role/page.tsx
@@ -59,28 +59,26 @@ const newAuthStore = newStoreDef<
   reducer({ prevState, state, action, diff, set, async, events, optimistic }) {
     if (action?.type === "change-role") {
       optimistic({ role: action.role })
-      async.promise(
-        async ({ signal }) => {
+      async()
+        .setName("role-request")
+        .promise(async ({ signal }) => {
           const role = await fetchRole({ roleName: action.role, signal })
           events.emit("got-role", role)
           set({ role })
-        },
-        { label: "role request" }
-      )
+        })
     }
 
     if (prevState.role !== state.role) {
-      async.promise(
-        async ({ signal }) => {
+      async()
+        .setName("fetch-permissions")
+        .promise(async ({ signal }) => {
           const permissions = await fetchPermissions({
             role: state.role,
             signal,
           })
           events.emit("got-permissions", permissions)
           set({ $permissions: permissions })
-        },
-        { label: "Fetch permissions" }
-      )
+        })
     }
 
     return state